refactor : constatns 파일 index.ts 하나로 모으기
