--- src/constants/errorMessage.ts
@@ -1,5 +1,0 @@
-export const ERROR_MESSAGE: { [key: string]: string } = {
-  400: "유효하지 않은 요청입니다. 입력한 정보를 다시 확인해주세요.",
-  404: "요청한 정보를 찾을 수 없습니다. ",
-  500: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요. ",
-};
