--- src/constants/options.ts
@@ -1,11 +1,0 @@
-export const categoryOptions = [
-  "전체",
-  "fashion",
-  "beverage",
-  "electronics",
-  "kitchen",
-  "fitness",
-  "books",
-];
-
-export const sortOptions = ["낮은 가격순", "높은 가격순"];
