--- src/constants/page.ts
@@ -1,9 +1,0 @@
-export const PAGE = {
-  FIRST_PAGE: 0,
-  FIRST_PAGE_LIMIT: 20,
-  SECOND_PAGE_STD: 5,
-  NEXT_PAGE_STEP: 1,
-  OTHER_PAGE_LIMIT: 4,
-  FIRST_PAGE_START: 0,
-  OTHER_PAGE_START: 2,
-};
