--- src/constants/styleOption.ts
@@ -1,3 +1,0 @@
-export const STYLE_OPTION = {
-  HEADER_Z_INDEX: 10,
-};
